+++ Chapter_1/Excercise_1.1/hello_world.c
@@ -1,0 +1,12 @@
+#include <stdio.h>
+
+/*
+Run the hello, world program on your system. Experiment with leaving out parts of the program,
+to see what error messages you get.
+*/
+
+int main()
+{
+	printf("Hello World!");
+	return 0;
+}