+++ Chapter_1/Excercise_1.2/hello_world_with_arg.c
@@ -1,0 +1,12 @@
+#include <stdio.h>
+
+/*
+Experiment to find out what happens when prints's argument string contains \c, where c is some
+character not listed above.
+*/
+
+int main()
+{
+    printf("Hello World!\c");
+	return 0;
+}