+++ Chapter_1/Excercise_1.3/fahr_to_celsius.c
@@ -1,5 +1,9 @@
 #include <stdio.h>
 
+/*
+Modify the temperature conversion program to print a heading above the table.
+*/
+
 int main()
 {
     float fahr, celsius;