+++ Chapter_1/Excercise_1.4/celsius_to_fahr.c
@@ -1,5 +1,9 @@
 #include <stdio.h>
 
+/*
+Write a program to print the corresponding Celsius to Fahrenheit table.
+*/
+
 int main()
 {
     float fahr, celsius;