+++ Chapter_1/Excercise_1.6/eof_expression.c
@@ -1,0 +1,11 @@
+#include <stdio.h>
+
+/*
+Verify the expression getchar() != EOF is 0 or 1.
+*/
+
+int main()
+{
+    printf("Value of expression: %d", getchar() != EOF);
+    return 0;
+}