+++ Chapter_1/Excercise_1.7/eof_printf.c
@@ -1,0 +1,11 @@
+#include <stdio.h>
+
+/*
+Write a Program to print the value of EOF.
+*/
+
+int main()
+{
+    printf("EOF: %d", EOF);
+    return 0;
+}